Move struct/fn into signature_with_sighash_info.rs file
